Create hourGlassSum.cpp
An hourglass in  is a subset of values with indices falling in this pattern in an array:
a b c
  d
e f g
An hourglass sum is the sum of an hourglass' values. Calculate the hourglass sum for every hourglass in the array, then print the maximum hourglass sum. The array will always be 6 x 6.